use int instead of long for block height (expiry)

+++ grpc-adapter/src/main/java/de/cotto/lndmanagej/grpc/GrpcPayments.java
@@ -50,7 +50,7 @@ public class GrpcPayments {
                     Instant expiryTimestamp = creationTimestamp.plusSeconds(payReq.getExpiry());
                     return new DecodedPaymentRequest(
                             paymentRequest,
-                            payReq.getCltvExpiry(),
+                            (int) payReq.getCltvExpiry(),
                             payReq.getDescription(),
                             Pubkey.create(payReq.getDestination()),
                             Coins.ofMilliSatoshis(payReq.getNumMsat()),
