+++ model/src/main/java/de/cotto/lndmanagej/model/DecodedPaymentRequest.java
@@ -4,7 +4,7 @@ import java.time.Instant;
 
 public record DecodedPaymentRequest(
         String paymentRequest,
-        long cltvExpiry,
+        int cltvExpiry,
         String description,
         Pubkey destination,
         Coins amount,
